Moved clients to glue.viewers
