+++ glue/clients/__init__.py
@@ -1,8 +1,3 @@
 from __future__ import absolute_import, division, print_function
 
-from .histogram_client import HistogramClient
-from .image_client import ImageClient
-from .scatter_client import ScatterClient
-
-from .viz_client import GenericMplClient
 from .layer_artist import LayerArtist
