+++ glue/viewers/histogram/client.py
@@ -9,10 +9,13 @@ from glue.core.data import Data
 from glue.core import message as msg
 from glue.core.client import Client
 from glue.core.roi import RangeROI
-from glue.clients.layer_artist import HistogramLayerArtist, LayerArtistContainer
+from glue.clients.layer_artist import LayerArtistContainer
 from glue.clients.util import update_ticks, visible_limits
-from glue.clients.viz_client import init_mpl
 from glue.utils import lookup_class
+
+from glue.viewers.common.viz_client import init_mpl
+
+from .layer_artist import HistogramLayerArtist
 
 
 class UpdateProperty(CallbackProperty):
