+++ glue/viewers/scatter/tests/test_layer_artist.py
@@ -1,9 +1,9 @@
 from __future__ import absolute_import, division, print_function
 
 from glue.core import Data
+from glue.utils import renderless_figure
 
 from ..layer_artist import ScatterLayerArtist
-from glue.utils import renderless_figure
 
 FIGURE = renderless_figure()
